+++ python/shotgun_view/widget_delegate.py
